@@ -27,10 +27,20 @@ namespace GroceryShop.Angular
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2)
-                // tratamento json loopping format
-                .AddJsonOptions(options => options.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore);
+            services.AddControllersWithViews();
+            // In production, the Angular files will be served from this directory
+            services.AddSpaStaticFiles(configuration =>
+            {
+                configuration.RootPath = "ClientApp/dist";
+            });
+
+            //services.AddMvc()
+            //    .SetCompatibilityVersion(CompatibilityVersion.Version_2_2)
+            //    // tratamento json loopping format
+            //    .AddJsonOptions(options => options.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore);
+
+            //services.AddMvc()
+               //.SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
 
             // String de conex�o com o Banco de dados (MySql)
             //var connectionString = Configuration.GetConnectionString("QuickByConnection");
@@ -52,28 +62,28 @@ namespace GroceryShop.Angular
 
 
             // Sweggar configuration
-            services.AddSwaggerGen(options =>
-            {
-                options.SwaggerDoc("v1",
-                    new Info
-                    {
-                        Title = "QuickBuy Angular APP",
-                        Version = "v1",
-                        Description = "Aplica��o Angular com ASP.Net Core",
-                        Contact = new Contact
-                        {
-                            Name = "Emerson Bilieri Claudelino",
-                            Url = "https://github.com/ebilieri"
-                        }
-                    });
-            });
+            //services.AddSwaggerGen(options =>
+            //{
+            //    options.SwaggerDoc("v1",
+            //        new Info
+            //        {
+            //            Title = "QuickBuy Angular APP",
+            //            Version = "v1",
+            //            Description = "Aplica��o Angular com ASP.Net Core",
+            //            Contact = new Contact
+            //            {
+            //                Name = "Emerson Bilieri Claudelino",
+            //                Url = "https://github.com/ebilieri"
+            //            }
+            //        });
+            //});
 
 
             // In production, the Angular files will be served from this directory
-            services.AddSpaStaticFiles(configuration =>
-            {
-                configuration.RootPath = "ClientApp/dist";
-            });
+            //services.AddSpaStaticFiles(configuration =>
+            //{
+            //    configuration.RootPath = "ClientApp/dist";
+            //});
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -92,23 +102,18 @@ namespace GroceryShop.Angular
 
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseSpaStaticFiles();
-
-            // Enable middleware to serve generated Swagger as a JSON endpoint.
-            app.UseSwagger();
-
-            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.), 
-            // specifying the Swagger JSON endpoint.
-            app.UseSwaggerUI(c =>
+            if (!env.IsDevelopment())
             {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "QuickBuy Angular V1");
-            });
+                app.UseSpaStaticFiles();
+            }
 
-            app.UseMvc(routes =>
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller}/{action=Index}/{id?}");
+                    pattern: "{controller}/{action=Index}/{id?}");
             });
 
             app.UseSpa(spa =>
@@ -120,14 +125,44 @@ namespace GroceryShop.Angular
 
                 if (env.IsDevelopment())
                 {
-                    /*
-                     * Comentar linha para executar o angular diretamente do angulur cli (ng build)
-                     * npm start - executar app
-                     */
                     spa.UseAngularCliServer(npmScript: "start");
-                    //spa.UseProxyToSpaDevelopmentServer("http://localhost:4200/");
                 }
             });
+
+            // Enable middleware to serve generated Swagger as a JSON endpoint.
+            //app.UseSwagger();
+
+            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.), 
+            // specifying the Swagger JSON endpoint.
+            //app.UseSwaggerUI(c =>
+            //{
+            //    c.SwaggerEndpoint("/swagger/v1/swagger.json", "QuickBuy Angular V1");
+            //});
+
+            //app.UseMvc(routes =>
+            //{
+            //    routes.MapRoute(
+            //        name: "default",
+            //        template: "{controller}/{action=Index}/{id?}");
+            //});
+
+            //app.UseSpa(spa =>
+            //{
+            //    // To learn more about options for serving an Angular SPA from ASP.NET Core,
+            //    // see https://go.microsoft.com/fwlink/?linkid=864501
+
+            //    spa.Options.SourcePath = "ClientApp";
+
+            //    if (env.IsDevelopment())
+            //    {
+            //        /*
+            //         * Comentar linha para executar o angular diretamente do angulur cli (ng build)
+            //         * npm start - executar app
+            //         */
+            //        spa.UseAngularCliServer(npmScript: "start");
+            //        //spa.UseProxyToSpaDevelopmentServer("http://localhost:4200/");
+            //    }
+            //});
         }
     }
 }
